Treat stopwords the same as unknown words #2484
Run tokens_by_line() first without a stopwords argument.
This allows stopword tokens to be included in the token stream and
later to be treated as "unknown" tokens. This way the presence of
stopwords in a match can impact a license match score and a license rule
with `with_only_known_words: yes` annot be matched not only if there
are unknown words but also if there are stopwords mixed its rule words.

Also add a simple end-to-end integration test for this.

Reported-by: Pierre Tardy <pierre.tardy@renault.com>
Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>